examples/skills-hello-world/pyproject.toml
[project]
name = "skills-hello-world"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "e2b-code-interpreter>=2.4.1",
    "tencentcloud-sdk-python>=3.0.0",
]
